# Switch to rle_decode_fast crate