--- a/Cargo.toml
+++ b/Cargo.toml
@@ -18,7 +18,7 @@ codecov = {repository = "sile/libflate"}
 adler32 = "1"
 byteorder = "1"
 crc32fast = "1"
-rle-decode-fast = {git = "https://github.com/Shnatsel/rle-decode-helper.git"}
+rle-decode-fast = "1.0.0"
 take_mut = "0.2.2"
 
 [dev-dependencies]