--- a/src/deflate/decode.rs
+++ b/src/deflate/decode.rs
@@ -3,6 +3,7 @@ use byteorder::ReadBytesExt;
 use std::cmp;
 use std::io;
 use std::io::Read;
+use rle_decode_helper::rle_decode;
 
 use super::symbol;
 use bit;
@@ -116,19 +117,7 @@ where
                             distance
                         ));
                     }
-                    let old_len = self.buffer.len();
-                    self.buffer.reserve(length as usize);
-                    unsafe {
-                        self.buffer.set_len(old_len + length as usize);
-                        let start = old_len - distance as usize;
-                        let ptr = self.buffer.as_mut_ptr();
-                        util::ptr_copy(
-                            ptr.add(start),
-                            ptr.add(old_len),
-                            length as usize,
-                            length > distance,
-                        );
-                    }
+                    rle_decode(&mut self.buffer, usize::from(distance), usize::from(length));
                 }
                 symbol::Symbol::EndOfBlock => {
                     break;