--- a/src/util.rs
+++ b/src/util.rs
@@ -2,17 +2,6 @@
 use std::io::{self, Read};
 use std::ptr;
 
-#[inline]
-pub unsafe fn ptr_copy(src: *const u8, dst: *mut u8, count: usize, is_overlapping: bool) {
-    if !is_overlapping {
-        ptr::copy_nonoverlapping(src, dst, count);
-    } else {
-        for i in 0..count {
-            ptr::copy_nonoverlapping(src.add(i), dst.add(i), 1);
-        }
-    }
-}
-
 #[cfg(test)]
 pub struct WouldBlockReader<R> {
     inner: R,