Use `rle_decode_fast` crate created by @WanzenBug and me instead of a bespoke unsafe block relying on non-local invariants.

This change improves performance by about 10% as measured with `hyperfine -m 25 --warmup=3 'target/release/examples/flate -i enwiki-latest-all-titles-in-ns0.gz -o /dev/null gzip-decode'`

Before:
```
  Time (mean ± σ):      2.295 s ±  0.016 s    [User: 2.275 s, System: 0.013 s]
  Range (min … max):    2.268 s …  2.320 s    25 runs
```

After:
```
  Time (mean ± σ):      2.079 s ±  0.022 s    [User: 2.069 s, System: 0.010 s]
  Range (min … max):    2.046 s …  2.132 s    25 runs
```


We're created the crate to solve recurring safety problems in RLE decoding. It contains a single unsafe block, which is a copy of `append_from_within()` function from [this Rust RFC](https://github.com/rust-lang/rfcs/pull/2714). We're reasonably confident it is sound because it's mostly a copy of a pre-exisitng stdlib function, and I cannot find anything wrong with its logic.